+++ pyespn/core/schedule.py
@@ -7,7 +7,6 @@ if TYPE_CHECKING:
     from pyespn.classes import Schedule
 
 
-# todo 1 is preseason 2 is regular season and 3 is postseason
 def get_regular_season_schedule_core(league_abbv, espn_instance,
                                      season, load_odds: bool = False,
                                      load_pbp: bool = False,
@@ -25,6 +24,8 @@ def get_regular_season_schedule_core(league_abbv, espn_instance,
             - '1' = preseason
             - '2' = regular season (default)
             - '3' = postseason
+            - '4' = offseason
+            - '5' = playin
 
     Returns:
         Schedule: A `Schedule` object containing the schedule for the specified season and league.
@@ -46,7 +47,7 @@ def get_regular_season_schedule_core(league_abbv, espn_instance,
     pages = content.get('pageCount')
     weeks_urls = []
     for page in range(1, pages + 1):
-        url = f'http://sports.core.api.espn.com/{v}/sports/{api_info["sport"]}/leagues/{api_info["league"]}/seasons/{season}/types/2/weeks?page={page}'
+        url = f'http://sports.core.api.espn.com/{v}/sports/{api_info["sport"]}/leagues/{api_info["league"]}/seasons/{season}/types/{season_type}/weeks?page={page}'
         page_content = fetch_espn_data(url)
         for item in page_content.get('items', []):
             weeks_urls.append(item.get('$ref'))