+++ pyespn/data/data_import.py
@@ -56,5 +56,8 @@ nascar_teams_data = open_json(team_lookup_file)
 team_lookup_file = 'files/indy_teams_lookup.json'
 indy_teams_data = open_json(team_lookup_file)
 
-team_lookup_file = 'files/epl_teams_ints.json'
+team_lookup_file = 'files/epl_teams_lookup.json'
 epl_teams_data = open_json(team_lookup_file)
+
+team_lookup_file = 'files/nhl_teams_lookup.json'
+nhl_teams_data = open_json(team_lookup_file)