+++ pyespn/data/teams.py
@@ -12,7 +12,8 @@ LEAGUE_TEAMS_MAPPING = {
     'f1': f1_teams_data,
     'nascar': nascar_teams_data,
     'indy': indy_teams_data,
-    'epl': epl_teams_data
+    'epl': epl_teams_data,
+    'nhl': nhl_teams_data
 }
 
 team_files_path = './files/'